data-API/requirements.txt
kafka-python
pytz
python-dateutil